edge_processor/requirements.txt
paho-mqtt
pandas
scikit-learn
xgboost==2.1.4
